BUG: Fix decimal search filter not showing up
Also refactor `TextField` to `InputField` so that `NumberField` can extend it.
