+++ client/src/components/InputField/InputField.js
@@ -1,0 +1,88 @@
+import React, { Component } from 'react';
+import { Input } from 'reactstrap';
+import PropTypes from 'prop-types';
+
+class InputField extends Component {
+  constructor(props) {
+    super(props);
+
+    this.handleChange = this.handleChange.bind(this);
+  }
+
+  /**
+   * Fetches the properties for the input field
+   *
+   * @returns {object} properties
+   */
+  getInputProps() {
+    const props = {
+      className: `${this.props.className} ${this.props.extraClass}`,
+      id: this.props.id,
+      name: this.props.name,
+      disabled: this.props.disabled,
+      readOnly: this.props.readOnly,
+      value: this.props.value || '',
+      placeholder: this.props.placeholder,
+      autoFocus: this.props.autoFocus,
+      maxLength: this.props.data && this.props.data.maxlength,
+      type: this.props.type ? this.props.type : null,
+    };
+
+    if (this.props.attributes && !Array.isArray(this.props.attributes)) {
+      Object.assign(props, this.props.attributes);
+    }
+
+    if (!this.props.readOnly) {
+      Object.assign(props, {
+        onChange: this.handleChange,
+      });
+    }
+
+    return props;
+  }
+
+  /**
+   * Handles changes to the input field's value.
+   *
+   * @param {Event} event
+   */
+  handleChange(event) {
+    if (typeof this.props.onChange === 'function') {
+      if (!event.target) {
+        return;
+      }
+      this.props.onChange(event, { id: this.props.id, value: event.target.value });
+    }
+  }
+
+  render() {
+    return <Input {...this.getInputProps()} />;
+  }
+}
+
+InputField.propTypes = {
+  extraClass: PropTypes.string,
+  id: PropTypes.string,
+  name: PropTypes.string.isRequired,
+  onChange: PropTypes.func,
+  value: PropTypes.oneOfType([PropTypes.string, PropTypes.number]),
+  readOnly: PropTypes.bool,
+  disabled: PropTypes.bool,
+  placeholder: PropTypes.string,
+  type: PropTypes.string,
+  autoFocus: PropTypes.bool,
+  attributes: PropTypes.oneOfType([PropTypes.object, PropTypes.array]),
+};
+
+InputField.defaultProps = {
+  // React considers "undefined" as an uncontrolled component.
+  value: '',
+  extraClass: '',
+  className: '',
+  type: 'text',
+  attributes: {},
+};
+
+export { InputField as Component };
+
+export default InputField;
